requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
pandas==2.2.3
greenlet==3.1.1
mysql-connector-python==9.2.0
numpy==1.23.5
parso==0.8.4
pip==25.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
gunicorn
azure-identity
azure-keyvault-secrets
python-dotenv
